dyninst: fix mapping of compile units for dwarf v5
The mapping of compile units to dwarf versions was trying to correct for
an alleged Delve off-by-one bug, but that bug does not seem to be real
-- I've checked by hand that the compile unit offsets we get for Delve
are correct. It would appear that Delve is correct for the released
version of Go 1.25; perhaps it wasn't correct for a pre-release version.

diff --git c/pkg/dyninst/object/elf.go i/pkg/dyninst/object/elf.go
new file mode 100644
index 0000000000..e69de29bb2