dynist/decoder: resolve unknown interface type name
This does plumbing to resolve unknown interface type names to the name
inside the runtime. This isn't perfect; unlike elsewhere, we're using
the Go runtime's name for the type, which only has a package name prefix
as opposed to a fully qualified name.

The expansion of this name is left for future work.