Implement ddot-BYOC workflow tool for deb and rpm packages (#39978)
### What does this PR do?

Extends the existing `Dockerfile.agent-otel` used for providing BYOC on
containers (as documented
[here](https://docs.datadoghq.com/opentelemetry/setup/ddot_collector/custom_components/))
to make it able to produce Linux _deb_ and _rpm_ files.

### Motivation

[ABLD-102](https://datadoghq.atlassian.net/browse/ABLD-102) (and
proposal described on that ticket).

### Describe how you validated your changes

I validated this locally with some basic smoke testing, and added a
couple of build jobs to CI to confirm that the artifacts are built (this
is not very thorough testing but it confirms the mechanism essentially
works).

### Possible Drawbacks / Trade-offs

- Most of these are currently documented on the proposal doc linked in
the Jira card.
- The `--output` option (and possibly some of the things added to the
Dockerfile) require buildkit. I don't see a good reason why a user with
access to docker would have a problem with this.
- Using `nfpm` instead of omnibus here means less opportunities to reuse
certain things (notably around version specification), but trying to
make this fit into omnibus looks like more work than it's worth.

### Additional Notes

The expected way to use this is visible in the added CI jobs, looking
like this:

```
      docker build \
        --target artifact \
        --output type=local,dest=./ \
        --build-arg AGENT_BRANCH=$CI_COMMIT_REF_NAME \
        --build-arg PACKAGE_TYPE=rpm \
        --build-arg AGENT_VERSION=$AGENT_VERSION \
        -f Dockerfile.agent-otel .
```

There are two main pieces of work left for follow-up PRs:
- Adding dependency related metadata to the packages to have it match
what the regular `ddot` package does as well as to make it conflict with
ddot itself.
- Support for selecting the base ubuntu version for the builder, such
that the compatibility for the resulting binaries can be adjusted.

[ABLD-102]:
https://datadoghq.atlassian.net/browse/ABLD-102?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Nicolas Schweitzer <nicolas.schweitzer@datadoghq.com>