[AGENTRUN-627] Add check metric to COAT (#39616)
<!--  
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:  
  * Should only fix one issue or add one feature at a time.  
  * Must update the test suite for the relevant functionality.  
  * Should pass all status checks before being reviewed or merged.  
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->  

### What does this PR do?  
Add the `checks.delay` and `checks.runs` telemetry metrics in the
default Agent telemetry profiles configuration. With tags: `check_name`
and `state`.

### Motivation  
To enhance the granularity of telemetry data for check execution by
enabling aggregation on key tags. This allows better filtering and
analysis of check performance and outcomes.

### Describe how you validated your changes  
Validated the update by building the Agent and confirming the new
metrics are registered with `sudo datadog-agent diagnose show-metadata
agent-telemetry`.

### Possible Drawbacks / Trade-offs  

### Additional Notes  
None.

---------

Co-authored-by: May Lee <may.lee@datadoghq.com>