[workloadmeta/kubelet] Stop relying on the pod watcher (#40099)
### What does this PR do?

This PR changes the kubelet workloadmeta collector to query the Kubelet
directly instead of using the PodWatcher.

The PodWatcher has caused several issues in the past. For example,
subtle bugs like missing tags under specific conditions. Some of the
reasons are:
- Custom update logic: It uses its own logic to detect when containers
or pods have changed, which differs from workloadmeta's approach
(`reflect.DeepEqual`).
- Custom "ready" logic: It overrides the pod's ready field with its own
logic.
- Automatic deletion of containers: It deletes containers of pods that
went unready (by its own definition) after 25 seconds, even when the
Kubelet still returns them.

With the changes in this PR, instead of relying on the PodWatcher, the
kubelet workloadmeta collector now:
- Queries the Kubelet directly.
- Uses workloadmeta's standard logic to detect updates.
- Reports exactly what the Kubelet returns without custom modifications.

Any custom logic that was in the PodWatcher should be moved to the
specific components that need it, not applied globally in workloadmeta
where it affects all subscribers. As an example, this PR moves the
"tolerate-unready" annotation logic from the PodWatcher to the
autodiscovery component where it belongs.

To reduce risk, this PR adds a new configuration option to fallback to
the old behavior. Set `kubelet_use_pod_watcher: true` to use the pod
watcher.

Workloadmeta was the only user of the PodWatcher. The other one was the
KSM check, but it was already migrated in a previous PR:
https://github.com/DataDog/datadog-agent/pull/39657

### Describe how you validated your changes

Unit tests for the new code plus existing e2e tests.
Also, manually verified that some of the components that could be
impacted still work as expected: autodiscovery, logs, etc.