fix(serverless-init): set serverless log hostname to empty string (#40074)
<!--
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->
### What does this PR do?

Passes the `hostnameinterface` component to the Serverless Logs Agent so
the hostname on logs is set to an empty string.

### Motivation

https://datadoghq.atlassian.net/browse/SVLS-7460

Currently the hostname is being set to `unknown`. In Datadog Orgs where
a host with the name `unknown` already exists and has tags on it, logs
with a matching hostname will inherit these tags, potentially resulting
in scenarios where a log has multiple tags with the same key but
different values.

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes.
If you validated changes manually and not through automated tests, add
context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask
reviewers to do it. Describe how to set up an environment for manual
tests in the PR description. Manual validation is expected to happen on
every commit before merge.

Any manual validation step should then map to an automated test. Manual
validation should not substitute automation, minus exceptions not
supported by test tooling yet.
-->

Added a serverless-init log test to validate that the hostname component
has an empty string set for a build of serverless-init.

Also deployed manually to Google Cloud Run and Azure Container Apps
using both the sidecar and in process approach to validate the logs have
hostname set as an empty string.

### Possible Drawbacks / Trade-offs

### Additional Notes

A regression was introduced in `v1.2.1` of the serverless-init image
that caused the hostname to be set to `unknown`. It's unclear what
specific change caused the regression as no Github tags were created for
these older versions.

There is already a provider hostname configuration for Serverless. By
passing the `hostnameinterface` component to the Serverless Logs Agent
we can set the empty string hostname on the provider.

https://github.com/DataDog/datadog-agent/blob/cc85b5170958191c3385c517cc9d9763654fdb12/pkg/util/hostname/providers_serverless.go#L19

Ultimately the hostname is passed to the logs processor where it is read
and inserted into the log message. This is where the hostname was set to
`unknown` prior to this change.

https://github.com/DataDog/datadog-agent/blob/35ec44794e8b6f90d8fb5c5c208b086f62dadaa0/pkg/logs/processor/processor.go#L199-L219

Since the signature of `SetupLogAgent` is updated to include the
hostname component `cmd/serverless` needs to be updated as well. However
Lambda Functions have special handling in the logs processor which means
this change will not affect the hostname on Lambda Function logs.

https://github.com/DataDog/datadog-agent/blob/42a25bbcf8055ea5c9be056f5bf539754d147df7/cmd/serverless-init/log/log.go#L59