Fix queue index check for networkv2 code (#40122)
The network check code looks for metrics like queue_0_tx_len and will
parse the queue index in order to set the tags correctly. However,
sometimes there are also metrics that look like tx_queue_drops which
contain the word "queue" but arent' actually queue-indexed metrics.

The old python checks would detect this case and stop trying to parse
the queue number. The networkv2 go checks do not detect this case and
would blow up instead.
(https://github.com/DataDog/integrations-core/blob/master/network/datadog_checks/network/ethtool.py#L106-L107)

This change implements the check, avoiding a crash in the case described
above.

### Describe how you validated your changes
I wrote a unit test which failed, then I changed the code so that it
passed.

No e2e testing (I'm still figuring that env out) because this seems like
the sort of code best validated through unit tests and there was a
pre-existing unit test for this module.

### Possible Drawbacks / Trade-offs
My biggest concern about this change is that I'm classifying the metrics
correctly. If the metric name fails to parse as a queue, it will fall
through and probably hit the generic final clause which checks to see if
this metric name is in the list of known metrics. I hope that's the
right behavior.

I would not expect any performance changes in either CPU, Mem usage, or
Disk usage as a result of this change.