dyninst/object: clean up concepts and API (#40124)
Before this change, there were some sprawling and poorly named
abstractions. The ElfFile exposed an MMappingElfFile but was really
about DWARF. There was also no way to get access to the object and its
dwarf without a file path and mmapping. This was unfortunate.

Now we have two clear interfaces: `File` and `Dwarf` (and
`FileWithDwarf`). There are now 4 implementations:

 * `InMemoryElfFile`: everything in RAM, decompresses on-demand to RAM.
 * `MMappingElfFile`: mmaps sections, never decompresses.
* `ElfFileWithDwarf`: decompresses data either to RAM or temporary files
on disk depending on how it is loaded.
* `InMemoryElfFileWithDwarf`: primarily added for testing, thin wrapper
around safeelf

This will enable testing that doesn't need disk space.

Relates to
[DEBUG-4381](https://datadoghq.atlassian.net/browse/DEBUG-4381?atlOrigin=eyJpIjoiNjdiNDY2ZjVhZmFkNGY5OWFkODY2OTgwZmE3MjUzOWMiLCJwIjoiamlyYS1zbGFjay1pbnQifQ)

[DEBUG-4381]:
https://datadoghq.atlassian.net/browse/DEBUG-4381?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ