Resolve `git` hash through `subprocess.check_output`
This is a follow-up of #39622, where `invoke.Context.run` happens to
return an empty `stdout` when it should not, defeating the resolution
and leading to `Omnibus cache mutation` alerts.

The present change gets back to `subprocess.check_output` since its
behavior is predictable without having to tweak its parameters, and fail
should an empty string be returned.