[SBOM] store SBOM compressed in the workloadmeta (#40043)
<!--
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->
### What does this PR do?

Currently we store full SBOMs for all container image metadata in the
workloadmeta store. This is used both by the flare provider (that
creates a `sbom.json` file in the flare) and also by the periodic
refresh that will send the SBOM for all images periodically.

Now SBOMs can be big (a few megabytes), and we expect them to get bigger
when we enable more scanners (like the apps and libraries scanners).
This is especially impactful on nodes with a lot of images thanks to
binpacking (internally we call this nodeless), and can result in agent
OOM kills.

To improve this state, this PR converts the workloadmeta store to store
SBOMs in a proto-encoded + compressed format. SBOMs are decompressed on
demand when they are needed (which should be pretty rare).

### Motivation

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes.
If you validated changes manually and not through automated tests, add
context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask
reviewers to do it. Describe how to set up an environment for manual
tests in the PR description. Manual validation is expected to happen on
every commit before merge.

Any manual validation step should then map to an automated test. Manual
validation should not substitute automation, minus exceptions not
supported by test tooling yet.
-->

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the
proposed
  version was chosen.
-->