Add commit sha to images pushed by `dev_branch-full` (#39465)
### What does this PR do?

Adds a new destination for `dev_branch-full` that includes
`CI_COMMIT_SHORT_SHA`

### Motivation

When pulling arbitrary agent images to do one off testing with SMP it is
often useful to manually run `dev_branch-full` to push an image from a
pipeline on main to docker/agent-dev but in order to more accurately
trace back to what commit we pulled from, having the commit sha would be
useful.

### Describe how you validated your changes

Will manually trigger

### Possible Drawbacks / Trade-offs

### Additional Notes

If there is a different way to do this I am open to suggestions. This is
very similar to this previous
[PR](https://github.com/DataDog/datadog-agent/pull/36203)

Signed-off-by: Caleb Metz <caleb.metz@datadoghq.com>