[SBOM] dump each container image SBOM into its own file in the flare (#40267)
<!--
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->
### What does this PR do?

When creating a flare, we currently create a big payload map mapping the
image ID (`sha256@123abc`) to the SBOM content as a json. This can be a
big payload, sometimes in the order of multiple dozens of MiB.

We recently started to compress the SBOM payload
(https://github.com/DataDog/datadog-agent/pull/40043) and the flare
creation is one of the last spots where we uncompress all SBOMs at the
same time.

To fix this, this PR changes the flare format to go from
```
sbom.json # with the big map as a json inside
```
to
```
sbom/
  sha256_123abc.json
  sha256_456def.json
```
with one file per image/SBOM. Allowing the creation to dump one file at
a time, and the GC to run if needed between decompression/serialization
of each file.

### Motivation

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes.
If you validated changes manually and not through automated tests, add
context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask
reviewers to do it. Describe how to set up an environment for manual
tests in the PR description. Manual validation is expected to happen on
every commit before merge.

Any manual validation step should then map to an automated test. Manual
validation should not substitute automation, minus exceptions not
supported by test tooling yet.
-->

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the
proposed
  version was chosen.
-->