Remove MockParams/MockModule from config mock
This way of initializing the config is no longer required and is yet
another way to init the config. To simplify the code and usage we're
removing it.

It also solves some broken tests where the order of initialization from
FX/non-FX mock could produce difference in the config between
initialization steps.