dyninst/symdb: properly synchronize uploads with RC state
Before this patch, we were uploading data anew every time we got an
"update" from the Scraper telling us that SymDB upload is requested; we
didn't keep track of the previous state of that flag so we were
uploading repeatedly. We also weren't responding to process removal
notifications from the procmon.

Now, we don't upload repeatedly, and we cancel an upload when the
process is removed.