dyninst/irgen: defer type creation
Before this change, we created types eagerly when we visited subprograms
and variables. This made it very hard to control the depth to which we
constructed the type graph. This refactor should have no functional
change, though it does rearrange the order in which we instantiate
types, and thus the type IDs.