dyninst/rcscrape: make debouncer dumber (#40292)
This is a minor refactor moving some of the responsibilities of the
debouncer to the parent Scraper. Now, the state of the debouncer is
simplified such that it only keeps track of "updates", without
explicitly caring any longer about procmon updates for processes coming
and going. Thus, it no longer needs to have state about tracked
processes that have no pending updates -- now it only knows about
processes that have pending updates.

The responsibility of dealing with procmon updates is lifted to the
scraper, which already had to deal with the procmon.