Allow older Ubuntu versions to be used as base for building ddot-byoc (#40218)
### What does this PR do?

It makes it possible to build custom ddot (byoc) targeting a broader
compatibility base (defined by available version of glibc on the OS), by
allowing the choice of the base version of Ubuntu used, via `--build-arg
UBUNTU_VERSION=xx.xx`.

### Motivation

Allow more users to benefit from the ddot byoc workflow.

### Describe how you validated your changes

Added a job to the pipeline which runs the build on Ubuntu 20.04 and
checks that the maximum glibc version required by the binary doesn't
exceed what would be expected for that version of Ubuntu.

### Possible Drawbacks / Trade-offs

- Not very user-friendly, but provides a quick way to provide this
compatibility.

### Additional Notes

This pins the `dda` version in use and switches to its "standalone"
version such that we don't need to worry about the python version or
python env management.

In practice, by dropping to Ubuntu 20.04, we currently already see the
glibc requirement drop to 2.17 max, which makes it at least as
compatible as the regular Agent.

We eventually plan to package our C/C++ toolchains that target the
compatibility that the Agent supports, which would probably give us a
better way to provide the necessary compatibility without the need to
change the base image like done here.