[Serverless-Init] Change Cloud Run Jobs origin tag; refactor logs source (#40276)
### What does this PR do?

- Renames the origin for cloud run jobs to `cloudrunjobs` (for billing
purposes)
- Adds the `origin:cloudrunjobs` tag to be used on the frontend
- Create a `GetDefaultLogsSource()` that will be used when creating the
agent log config, because we want Cloud Run Jobs to have
`origin:cloudrunjobs` but a logs source value of `cloudrun`
- Also, origin and source are not the same thing. The previous
implementation was confusing and seems to imply that the two values are
equivalent, when they are actually very different. This makes it clear
that we are just falling back to `GetDefaultLogsSource()` when
`DD_SOURCE` is not defined
- Although this is technically breaking (since we're changing the tag
value), this should not be considered breaking because Cloud Run Jobs is
not officially supported or released yet.

### Motivation

- Separate Cloud Run services and Cloud Run Jobs for billing purposes
- Expose the `origin` tag for the frontend

### Describe how you validated your changes

Built serverless-init locally and deployed it to a test Cloud Run Job.

As you can see, the tag is `origin:cloudrunjobs`, and the log source is
`cloudrun`:
<img width="500" height="349" alt="Screenshot 2025-08-27 at 1 36 51 PM"
src="https://github.com/user-attachments/assets/72c4b7ad-be0a-4913-8e3c-4ddb18b24fd7"
/>

### Possible Drawbacks / Trade-offs

None

### Additional Notes