Introduce dynamic index to run relevant tests (#39364)
<!--
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->
### What does this PR do?

Introduce dynamic index to execute tests as described in
https://datadoghq.atlassian.net/wiki/x/q4h2PgE
Create several components that allow to compute a dynamic index:
- **DynamicIndex**: Wrapper around a dict[str, dict[str, set[str]] that
represent an index giving a list of test to execute given some changes,
with some useful function to dump the index, merge it or load it from a
JSON.
- **DynTestBackend**: Represent a backend to store indexes with basic
crud operations , first implementation is using a S3 bucket with json
files stored.
- **DynTestIndexer**: Used to compute an index, first one implement is a
coverage-based indexer that use Go code coverage to compute the index
- **DynTestExecutor**: Used to retrieve the correct index and that given
a list of changes return the tests that must be executed.
- **DynTestEvaluator**: Not strictly needed to run the tests but useful
to evaluate how efficient the system is and to make sure it would not
hide issues.

More details can be found in the README of tasks/libs/dynamic_test and
in the docstrings.

Merging that PR will put in place a system to evaluate how efficient it
is to run e2e tests based on the coverage computed in a nightly
pipeline. A corresponding notebook was created to see the amount of test
executed with that new method vs the old way of doing it:
https://app.datadoghq.com/notebook/12935247/dynamic-test-execution-evaluation?range=172800000&start=1755600314643&live=true
It will also help to detect errors in the system and too optimistic test
skips.

Next steps is to let the system work for some days to assess how
efficient it is, to fix the eventual errors that are found. Once this is
working the executor should be used in the e2e tests directly to skip
the tests that are not needed.
In the meantime we should add coverage support for more tests, at that
moment only the host-based tests are able to compute their coverage.

Note: The evaluation is only made on tests that we have info about,
which means that all the tests that we are not collecting the coverage
yet are not included in the statistics.

The evaluate job compare test executed with that output:
https://gitlab.ddbuild.io/DataDog/datadog-agent/-/jobs/1102804231


### Motivation

The ultimate goal is make our CI more efficient by providing a way to
run only relevant tests on PR. It should reduce the cost of the CI and
the flakiness, should also reduce the number of failing tests that end
up in main because they were not executed on the PR

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes.
If you validated changes manually and not through automated tests, add
context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask
reviewers to do it. Describe how to set up an environment for manual
tests in the PR description. Manual validation is expected to happen on
every commit before merge.

Any manual validation step should then map to an automated test. Manual
validation should not substitute automation, minus exceptions not
supported by test tooling yet.
-->

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the
proposed
  version was chosen.
-->

Some `pip install boto3` are still done at runtime, they will be removed
as soon as test-infra-definitions is bumped to the latest version

---------

Co-authored-by: Célian Raimbault <161456554+CelianR@users.noreply.github.com>